NPC Sheet and localisation
- Added an NPC Sheet
- Added localisation. Full disclosure, the localised files were translated using an online tool so may be a bit wrong. If they are I am happy to accept corrections.
