+++ src/lang/de.json
@@ -1,0 +1,30 @@
+{
+	"character": {
+		"name": "Name",
+		"pronouns": "Pronomen",
+		"outcomedice": "Ergebniswürfel",
+		"luck": "Glück",
+		"organisation": "Organisation",
+		"background": "Hintergrund",
+		"role": "Rolle",
+		"niche": "Nische",
+		"core": "Kern",
+		"quirks": "Macken",
+		"traits": "Eigenschaften",
+		"description": "Beschreibung",
+		"consequences": "Konsequenzen",
+		"notes": "Notizen"
+	},
+	"application": {
+		"createitem": "Gegenstand erstellen",
+		"edititem": "Gegenstand bearbeiten",
+		"deleteitem": "Gegenstand löschen",
+		"deleteconfirm": "Bist du sicher, dass du diesen Gegenstand löschen möchtest?",
+		"yes": "Ja",
+		"no": "Nein",
+		"discworlddiceroller": "Discworld Würfelwerfer",
+		"roll": "Würfeln",
+		"rolling": "Würfeln...",
+		"close": "Schließen"
+	}		
+}
