+++ src/module/discroller.js
@@ -2,7 +2,7 @@ export class DiscRoller {
   static async Init(controls, html) {
     // Create the main dice roll button
     const diceRollbtn = $(`
-      <li class="scene-control discworld-roller" data-control="DiscRoller" title="Discworld Dice Roller">
+      <li class="scene-control discworld-roller" data-control="DiscRoller" title="${game.i18n.localize('application.discworlddiceroller')}">
           <i class="fa-regular fa-hand-holding-magic"></i>
           <ol class="nested-buttons sub-controls control-tools"></ol>
       </li>
@@ -17,19 +17,19 @@ export class DiscRoller {
 
   static async CreateDiceRoller(event) {
     const dialog = new Dialog({
-      title: "Discworld Dice Roller",
+      title: `${game.i18n.localize('application.discworlddiceroller')}`,
       content: `
         <form>
           <div class="form-group split-row">
             <div class="dice-roll-container">
               <div class="dice-roll-buttons">
-                <button type="button" class="roll-button" data-dice="d4">Roll d4</button>
-                <button type="button" class="roll-button" data-dice="d6">Roll d6</button>
-                <button type="button" class="roll-button" data-dice="d8">Roll d8</button>
-                <button type="button" class="roll-button" data-dice="d10">Roll d10</button>
-                <button type="button" class="roll-button" data-dice="d12">Roll d12</button>
-                <button type="button" class="roll-button" data-dice="d20">Roll d20</button>
-                <button type="button" class="roll-button" data-dice="d100">Roll d100</button>
+                <button type="button" class="roll-button" data-dice="d4">${game.i18n.localize('application.roll')} d4</button>
+                <button type="button" class="roll-button" data-dice="d6">${game.i18n.localize('application.roll')} d6</button>
+                <button type="button" class="roll-button" data-dice="d8">${game.i18n.localize('application.roll')} d8</button>
+                <button type="button" class="roll-button" data-dice="d10">${game.i18n.localize('application.roll')} d10</button>
+                <button type="button" class="roll-button" data-dice="d12">${game.i18n.localize('application.roll')} d12</button>
+                <button type="button" class="roll-button" data-dice="d20">${game.i18n.localize('application.roll')} d20</button>
+                <button type="button" class="roll-button" data-dice="d100">${game.i18n.localize('application.roll')} d100</button>
               </div>
             </div>
           </div>
@@ -38,7 +38,7 @@ export class DiscRoller {
       `,
       buttons: {
         close: {
-          label: "Close",
+          label: `${game.i18n.localize('application.close')}`,
           callback: () => {}
         }
       },
@@ -53,7 +53,7 @@ export class DiscRoller {
 
           roll.toMessage({
             speaker: ChatMessage.getSpeaker({ actor: game.user.character }),
-            flavor: `Rolling ${formula}`
+            flavor: `${game.i18n.localize('application.rolling')} ${formula}`
           });
         });
       },
