+++ src/templates/actors/npc.hbs
@@ -1,0 +1,51 @@
+<form class="discworld-actor">
+    <div class="field-group name-field">
+      <label for="name">{{localize 'character.name'}}:</label>
+      <input type="text" name="name" value="{{actor.name}}" class="name-input">
+    </div>
+
+  <div class="form-group split-row">
+    <div class="field-group role-field">
+      <label for="role">{{localize 'character.role'}}:</label>
+      <input type="text" name="system.role" value="{{actor.system.role}}">
+    </div>
+
+    <div class="field-group pronouns-field">
+      <label for="pronouns">{{localize 'character.pronouns'}}:</label>
+      <input type="text" name="system.pronouns" value="{{actor.system.pronouns}}" class="pronouns-input">
+    </div>
+  </div>
+
+  <div class="form-group split-row">
+    <div class="left-column">
+
+      <label for="core">{{localize 'character.traits'}}:</label>
+      <a class="control create" title="{{localize 'application.createitem'}}" data-type="trait">
+        <i class="fas fa-plus"></i>
+      </a>
+      {{#each items as |item id|}}
+        {{#if (eq item.type "trait")}}
+          <li class="row entry" data-item-id="{{item._id}}" data-item-type="{{item.type}}" data-item-value="{{item.name}}">
+            <h4 class="one-column-no-roll list-entry">{{item.name}}</h4>
+            <div class="control">
+              <a class="edit" title="{{localize 'application.edititem'}}"><i class="fas fa-edit"></i></a>
+              <a class="delete" title="{{localize 'application.deleteitem'}}"><i class="fas fa-trash"></i></a>
+            </div>
+          </li>
+        {{/if}}
+      {{/each}}
+
+    <div class="form-group-notes">
+      <label for="notes">{{localize 'character.notes'}}:</label>
+      <textarea name="system.notes">{{actor.system.notes}}</textarea>
+    </div>
+    </div>
+
+    <div class="right-column">
+      <div class="form-group image-field">
+        <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}"/>
+      </div>
+    </div>
+  </div>
+
+</form>
